+++ sds_data_manager/stacks/ialirt_ingest_lambda_stack.py
@@ -87,12 +87,11 @@ class IalirtIngestLambda(Stack):
             managed_policies=[
                 iam.ManagedPolicy.from_aws_managed_policy_name(
                     "service-role/AWSLambdaBasicExecutionRole"
-                )
+                ),
+                iam.ManagedPolicy.from_aws_managed_policy_name(
+                    "AmazonDynamoDBFullAccess"
+                ),
             ],
-        )
-
-        lambda_role.add_managed_policy(
-            iam.ManagedPolicy.from_aws_managed_policy_name("AmazonDynamoDBFullAccess")
         )
 
         lambda_role.add_to_policy(
