+++ tests/infrastructure/test_ialirt_ingest_lambda_stack.py
@@ -9,11 +9,11 @@ def populate_table(table):
     """Populate DynamoDB table."""
     items = [
         {
-            "reset_number#met": "0#2025-07-11T12:34:56Z",
+            "reset_number#met": "0#123",
             "packet_blob": b"binary_data_string",
         },
         {
-            "reset_number#met": "0#2025-07-12T12:34:57Z",
+            "reset_number#met": "0#124",
             "packet_blob": b"binary_data_string",
         },
     ]
@@ -25,9 +25,7 @@ def populate_table(table):
 
 def test_query_by_sct_vtcw(table, populate_table):
     """Test to query irregular packet length."""
-    response = table.query(
-        KeyConditionExpression=Key("reset_number#met").eq("0#2025-07-12T12:34:57Z")
-    )
+    response = table.query(KeyConditionExpression=Key("reset_number#met").eq("0#124"))
 
     items = response["Items"]
-    assert items[0]["reset_number#met"] == "0#2025-07-12T12:34:57Z"
+    assert items[0]["reset_number#met"] == "0#124"
