+++ tests/lambda_endpoints/test_batch_starter.py
@@ -26,8 +26,7 @@ from sds_data_manager.lambda_code.SDSCode.dependency_config import (
     downstream_dependents,
     upstream_dependents,
 )
-
-from .conftest import POSTGRES_AVAILABLE
+from tests.conftest import POSTGRES_AVAILABLE
 
 
 def _populate_dependency_table(session):
